requirements-server.txt
# apipool-server dependencies

# Core library
sqlalchemy>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Encryption
cryptography>=41.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
python-dotenv>=1.0.0

# HTTP client (SDK)
httpx>=0.25.0
